SSE Fix & first stab at read timeouts
1) The SSEInboundHandler was not passing any unknown messages through it.
2) The SSE encoder was putting a single new line instead of two to indicate end of message.
3) Introduced the read timeout functionality. TODO: Remove the timeout handler.
